+++ temba/campaigns/tests.py
@@ -1,5 +1,4 @@
-from datetime import datetime, timedelta
-from unittest.mock import patch
+from datetime import timedelta
 
 import pytz
 
@@ -8,12 +7,11 @@ from django.core.exceptions import ValidationError
 from django.urls import reverse
 from django.utils import timezone
 
-from temba.contacts.models import Contact, ContactField, ContactGroup, ImportTask
+from temba.contacts.models import ContactField, ContactGroup
 from temba.flows.models import Flow, FlowRevision
 from temba.msgs.models import Msg
 from temba.orgs.models import Language, Org
-from temba.tests import TembaTest, matchers, mock_mailroom
-from temba.utils import json
+from temba.tests import CRUDLTestMixin, TembaTest, matchers, mock_mailroom
 from temba.values.constants import Value
 
 from .models import Campaign, CampaignEvent, EventFire
@@ -981,7 +979,7 @@ class CampaignTest(TembaTest):
 
         # give contact a last seen on value
         self.farmer1.last_seen_on = timezone.now()
-        self.farmer1.save(update_fields=("last_seen_on",), handle_update=False)
+        self.farmer1.save(update_fields=("last_seen_on",))
 
         ev4 = EventFire.objects.create(event=event3, contact=self.farmer1, scheduled=trim_date, fired=trim_date)
         self.assertIsNotNone(ev4.get_relative_to_value())
@@ -1026,7 +1024,7 @@ class CampaignTest(TembaTest):
         # give contact a last seen on value
         now = timezone.now()
         self.farmer1.last_seen_on = now
-        self.farmer1.save(update_fields=("last_seen_on",), handle_update=False)
+        self.farmer1.save(update_fields=("last_seen_on",))
 
         expected_result = (now + timedelta(days=5)).replace(second=0, microsecond=0).astimezone(self.org.timezone)
         self.assertEqual(event.calculate_scheduled_fire(self.farmer1), expected_result)
@@ -1047,120 +1045,6 @@ class CampaignTest(TembaTest):
         exported = response.json()
 
         self.org.import_app(exported, self.admin)
-
-    def test_deleting_reimport_contact_groups(self):
-        with patch.object(timezone, "now", return_value=datetime(2020, 5, 1, 0, 0, 0, 0, pytz.UTC)):
-            campaign = Campaign.create(self.org, self.admin, "Planting Reminders", self.farmers)
-
-            # create a reminder for our first planting event
-            planting_reminder = CampaignEvent.create_flow_event(
-                self.org,
-                self.admin,
-                campaign,
-                relative_to=self.planting_date,
-                offset=3,
-                unit="D",
-                flow=self.reminder_flow,
-            )
-
-            self.assertEqual(0, EventFire.objects.all().count())
-            self.set_contact_field(self.farmer1, "planting_date", "10-05-2020 12:30:10", legacy_handle=True)
-            self.set_contact_field(self.farmer2, "planting_date", "15-05-2020 12:30:10", legacy_handle=True)
-
-            # now we have event fires accordingly
-            self.assertEqual(2, EventFire.objects.all().count())
-
-            # farmer one fire
-            scheduled = EventFire.objects.get(contact=self.farmer1, event=planting_reminder).scheduled
-            self.assertEqual("13-5-2020", "%s-%s-%s" % (scheduled.day, scheduled.month, scheduled.year))
-
-            # farmer two fire
-            scheduled = EventFire.objects.get(contact=self.farmer2, event=planting_reminder).scheduled
-            self.assertEqual("18-5-2020", "%s-%s-%s" % (scheduled.day, scheduled.month, scheduled.year))
-
-            # delete our farmers group
-            self.farmers.release()
-
-            # this should have removed all the event fires for that group
-            self.assertEqual(0, EventFire.objects.filter(event=planting_reminder).count())
-
-            # and our group is no longer active
-            self.assertFalse(campaign.group.is_active)
-
-            # now import the group again
-            filename = "farmers.csv"
-            extra_fields = [dict(key="planting_date", header="field: planting_date", label="Planting Date", type="D")]
-            import_params = dict(
-                org_id=self.org.id,
-                timezone=str(self.org.timezone),
-                extra_fields=extra_fields,
-                original_filename=filename,
-            )
-
-            task = ImportTask.objects.create(
-                created_by=self.admin,
-                modified_by=self.admin,
-                csv_file="test_imports/" + filename,
-                model_class="Contact",
-                import_params=json.dumps(import_params),
-                import_log="",
-                task_id="A",
-            )
-            Contact.import_csv(task, log=None)
-
-            # check that we have new planting dates
-            self.farmer1 = Contact.objects.get(pk=self.farmer1.pk)
-            self.farmer2 = Contact.objects.get(pk=self.farmer2.pk)
-
-            planting = self.farmer1.get_field_value(self.planting_date)
-            self.assertEqual("10-8-2020", "%s-%s-%s" % (planting.day, planting.month, planting.year))
-
-            planting = self.farmer2.get_field_value(self.planting_date)
-            self.assertEqual("15-8-2020", "%s-%s-%s" % (planting.day, planting.month, planting.year))
-
-            # now update the campaign
-            new_farmers = ContactGroup.user_groups.filter(name="Farmers", is_active=True).first()
-            new_campaign = Campaign.create(self.org, self.admin, "Planting Reminders", new_farmers)
-            new_planting_reminder = CampaignEvent.create_flow_event(
-                self.org,
-                self.admin,
-                new_campaign,
-                relative_to=self.planting_date,
-                offset=3,
-                unit="D",
-                flow=self.reminder_flow,
-            )
-
-            self.login(self.admin)
-            post_data = dict(name="Planting Reminders", group=new_farmers.pk)
-
-            self.client.post(reverse("campaigns.campaign_update", args=[new_campaign.pk]), post_data)
-
-            self.set_contact_field(self.farmer1, "planting_date", "13-08-2020 12:30:10", legacy_handle=True)
-            self.set_contact_field(self.farmer2, "planting_date", "18-08-2020 12:30:10", legacy_handle=True)
-
-            # should have two fresh new fires
-            self.assertEqual(2, EventFire.objects.all().count())
-
-            # check their new planting dates
-            scheduled = EventFire.objects.get(contact=self.farmer1, event=new_planting_reminder).scheduled
-            self.assertEqual("16-8-2020", "%s-%s-%s" % (scheduled.day, scheduled.month, scheduled.year))
-
-            # farmer two fire
-            scheduled = EventFire.objects.get(contact=self.farmer2, event=new_planting_reminder).scheduled
-            self.assertEqual("21-8-2020", "%s-%s-%s" % (scheduled.day, scheduled.month, scheduled.year))
-
-            # give our non farmer a planting date
-            self.set_contact_field(self.nonfarmer, "planting_date", "20-05-2020 12:30:10", legacy_handle=True)
-
-            # now update to the non-farmer group
-            self.nonfarmers = self.create_group("Not Farmers", [self.nonfarmer])
-            post_data = dict(name="Planting Reminders", group=self.nonfarmers.pk)
-            self.client.post(reverse("campaigns.campaign_update", args=[new_campaign.pk]), post_data)
-
-            # which will cause event to be cloned again
-            new_planting_reminder.refresh_from_db()
-            self.assertFalse(new_planting_reminder.is_active)
 
     def test_update_to_non_date(self):
         # create our campaign and event
@@ -1617,43 +1501,34 @@ class CampaignTest(TembaTest):
         self.assertEqual(campaign_event.delivery_hour, -1)
 
 
-class CampaignCRUDLTest(TembaTest):
-    def setUp(self):
-        super().setUp()
-
-        self.campaign1 = self.create_campaign(self.org)
-        self.other_org_campaign = self.create_campaign(self.org2)
-
-    def create_campaign(self, org):
+class CampaignCRUDLTest(TembaTest, CRUDLTestMixin):
+    def create_campaign(self, org, name, group):
         user = org.get_user()
-        group = self.create_group("Reporters", contacts=[], org=org)
         registered = self.create_field("registered", "Registered", value_type="D", org=org)
         flow = self.create_flow(org=org)
-        campaign = Campaign.create(org, user, "Welcomes", group)
+        campaign = Campaign.create(org, user, name, group)
         CampaignEvent.create_flow_event(
             org, user, campaign, registered, offset=1, unit="W", flow=flow, delivery_hour="13"
         )
         return campaign
 
     def test_read(self):
-        read_url = reverse("campaigns.campaign_read", args=[self.campaign1.id])
+        group = self.create_group("Reporters", contacts=[])
+        campaign = self.create_campaign(self.org, "Welcomes", group)
 
-        # can't view campaign if not logged in
-        response = self.client.get(read_url)
-        self.assertLoginRedirect(response)
+        read_url = reverse("campaigns.campaign_read", args=[campaign.id])
 
-        self.login(self.admin)
-
-        response = self.client.get(read_url)
+        response = self.assertReadFetch(read_url, allow_viewers=True, allow_editors=True, context_object=campaign)
         self.assertContains(response, "Welcomes")
         self.assertContains(response, "Registered")
 
-        # can't view campaign from other org
-        response = self.client.get(reverse("campaigns.campaign_read", args=[self.other_org_campaign.id]))
-        self.assertLoginRedirect(response)
-
     def test_archive_and_activate(self):
-        archive_url = reverse("campaigns.campaign_archive", args=[self.campaign1.id])
+        group = self.create_group("Reporters", contacts=[])
+        campaign = self.create_campaign(self.org, "Welcomes", group)
+        other_org_group = self.create_group("Reporters", contacts=[], org=self.org2)
+        other_org_campaign = self.create_campaign(self.org2, "Welcomes", other_org_group)
+
+        archive_url = reverse("campaigns.campaign_archive", args=[campaign.id])
 
         # can't archive campaign if not logged in
         response = self.client.post(archive_url)
@@ -1664,23 +1539,84 @@ class CampaignCRUDLTest(TembaTest):
         response = self.client.post(archive_url)
         self.assertEqual(302, response.status_code)
 
-        self.campaign1.refresh_from_db()
-        self.assertTrue(self.campaign1.is_archived)
+        campaign.refresh_from_db()
+        self.assertTrue(campaign.is_archived)
 
         # activate that archve
-        response = self.client.post(reverse("campaigns.campaign_activate", args=[self.campaign1.id]))
+        response = self.client.post(reverse("campaigns.campaign_activate", args=[campaign.id]))
         self.assertEqual(302, response.status_code)
 
-        self.campaign1.refresh_from_db()
-        self.assertFalse(self.campaign1.is_archived)
+        campaign.refresh_from_db()
+        self.assertFalse(campaign.is_archived)
 
         # can't archive campaign from other org
-        response = self.client.post(reverse("campaigns.campaign_archive", args=[self.other_org_campaign.id]))
+        response = self.client.post(reverse("campaigns.campaign_archive", args=[other_org_campaign.id]))
         self.assertEqual(404, response.status_code)
 
         # check object is unchanged
-        self.other_org_campaign.refresh_from_db()
-        self.assertFalse(self.other_org_campaign.is_archived)
+        other_org_campaign.refresh_from_db()
+        self.assertFalse(other_org_campaign.is_archived)
+
+    @mock_mailroom
+    def test_update(self, mr_mocks):
+        group1 = self.create_group("Reporters", contacts=[])
+        group2 = self.create_group("Testers", contacts=[])
+        campaign = self.create_campaign(self.org, "Welcomes", group1)
+
+        update_url = reverse("campaigns.campaign_update", args=[campaign.id])
+
+        self.assertUpdateFetch(update_url, allow_viewers=False, allow_editors=True, form_fields=["name", "group"])
+
+        # try to submit with empty name
+        self.assertUpdateSubmit(
+            update_url,
+            {"name": "", "group": group1.id},
+            form_errors={"name": "This field is required."},
+            object_unchanged=campaign,
+        )
+
+        # submit with valid name
+        self.assertUpdateSubmit(update_url, {"name": "Greetings", "group": group1.id}, success_status=200)
+
+        campaign.refresh_from_db()
+        self.assertEqual("Greetings", campaign.name)
+        self.assertEqual(group1, campaign.group)
+
+        # group didn't change so shouldn't have queued any tasks to reschedule events
+        self.assertEqual(0, len(mr_mocks.queued_batch_tasks))
+
+        # submit with group change
+        self.assertUpdateSubmit(update_url, {"name": "Greetings", "group": group2.id}, success_status=200)
+
+        campaign.refresh_from_db()
+        self.assertEqual("Greetings", campaign.name)
+        self.assertEqual(group2, campaign.group)
+
+        # should have a task queued to reschedule the campaign's event
+        self.assertEqual(1, len(mr_mocks.queued_batch_tasks))
+        self.assertEqual(
+            {
+                "type": "schedule_campaign_event",
+                "org_id": self.org.id,
+                "task": {"campaign_event_id": campaign.events.filter(is_active=True).get().id, "org_id": self.org.id},
+                "queued_on": matchers.Datetime(),
+            },
+            mr_mocks.queued_batch_tasks[0],
+        )
+
+    def test_list(self):
+        group = self.create_group("Reporters", contacts=[])
+        campaign1 = self.create_campaign(self.org, "Welcomes", group)
+        campaign2 = self.create_campaign(self.org, "Follow Ups", group)
+
+        other_org_group = self.create_group("Reporters", contacts=[], org=self.org2)
+        self.create_campaign(self.org2, "Welcomes", other_org_group)
+
+        update_url = reverse("campaigns.campaign_list")
+
+        self.assertListFetch(
+            update_url, allow_viewers=True, allow_editors=True, context_objects=[campaign2, campaign1]
+        )
 
 
 class CampaignEventCRUDLTest(TembaTest):